Reorganizing the services and adding XML comments
