+++ AllReadyApp/Web-App/AllReady.UnitTest/Hangfire/Jobs/ProcessApiRequestsShould.cs
@@ -4,6 +4,7 @@ using AllReady.Features.Requests;
 using AllReady.Hangfire.Jobs;
 using AllReady.Models;
 using AllReady.Services.Mapping;
+using AllReady.Services.Mapping.GeoCoding;
 using AllReady.ViewModels.Requests;
 using MediatR;
 using Moq;
