+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Requests/EditRequestCommandHandler.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using AllReady.Extensions;
 using AllReady.Services.Mapping;
+using AllReady.Services.Mapping.GeoCoding;
 
 namespace AllReady.Areas.Admin.Features.Requests
 {
