+++ AllReadyApp/Web-App/AllReady/Services/Mapping/GeoCoding/Models/Google/GoogleGeocodeResult.cs
@@ -1,0 +1,10 @@
+﻿namespace AllReady.Services.Mapping.GeoCoding.Models.Google
+{
+    /// <summary>
+    /// Maps properties from a Google Geocode API call JSON response
+    /// </summary>
+    public class GoogleGeocodeResult
+    {
+        public GoogleGeocodeResultGeometry Geometry { get; set; }
+    }
+}
