+++ AllReadyApp/Web-App/AllReady/Services/Mapping/GeoCoding/Models/Google/GoogleGeocodeResultGeometry.cs
@@ -1,0 +1,10 @@
+﻿namespace AllReady.Services.Mapping.GeoCoding.Models.Google
+{
+    /// <summary>
+    /// Maps properties from a Google Geocode API call JSON response
+    /// </summary>
+    public class GoogleGeocodeResultGeometry
+    {
+        public GoogleGeocodeResultLocation Location { get; set; }
+    }
+}
