+++ AllReadyApp/Web-App/AllReady/Services/Mapping/GeoCoding/Models/Google/GoogleGeocodeResultLocation.cs
@@ -1,7 +1,10 @@
 ﻿using Newtonsoft.Json;
 
-namespace AllReady.Services.Mapping
+namespace AllReady.Services.Mapping.GeoCoding.Models.Google
 {
+    /// <summary>
+    /// Maps properties from a Google Geocode API call JSON response
+    /// </summary>
     public class GoogleGeocodeResultLocation
     {
         [JsonProperty("lat")]
