--- AllReadyApp/Web-App/AllReady/Services/Mapping/GoogleGeocodeResult.cs
@@ -1,7 +1,0 @@
-﻿namespace AllReady.Services.Mapping
-{
-    public class GoogleGeocodeResult
-    {
-        public GoogleGeocodeResultGeometry Geometry { get; set; }
-    }
-}
