--- AllReadyApp/Web-App/AllReady/Services/Mapping/GoogleGeocodeResultGeometry.cs
@@ -1,7 +1,0 @@
-﻿namespace AllReady.Services.Mapping
-{
-    public class GoogleGeocodeResultGeometry
-    {
-        public GoogleGeocodeResultLocation Location { get; set; }
-    }
-}
