+++ AllReadyApp/Web-App/AllReady/Services/Mapping/Routing/IOptimizeRouteService.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+
+namespace AllReady.Services.Mapping.Routing
+{
+    /// <summary>
+    /// A service which provides route optimization
+    /// </summary>
+    public interface IOptimizeRouteService
+    {
+        /// <summary>
+        /// Optimize a set of route criteria / waypoints
+        /// </summary>
+        Task<OptimizeRouteResult> OptimizeRoute(OptimizeRouteCriteria criteria);
+    }
+}
