+++ AllReadyApp/Web-App/AllReady/Services/Mapping/Routing/Models/Google/GoogleDirectionsResponse.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Newtonsoft.Json;
 
-namespace AllReady.Services.Routing.Models.Google
+namespace AllReady.Services.Mapping.Routing.Models.Google
 {
     public class GoogleDirectionsResponse
     {
