+++ AllReadyApp/Web-App/AllReady/Services/Mapping/Routing/Models/Google/GoogleDirectionsRouteLeg.cs
@@ -1,4 +1,4 @@
-﻿namespace AllReady.Services.Routing.Models.Google
+﻿namespace AllReady.Services.Mapping.Routing.Models.Google
 {
     public class GoogleDirectionsRouteLeg
     {
