+++ AllReadyApp/Web-App/AllReady/Services/Mapping/Routing/OptimizeRouteCriteria.cs
@@ -3,10 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.Encodings.Web;
 
-namespace AllReady.Services.Routing
+namespace AllReady.Services.Mapping.Routing
 {
+    /// <summary>
+    /// Defines the properties required as input for route optimization
+    /// </summary>
     public class OptimizeRouteCriteria
     {
+        /// <summary>
+        /// Initializes a new instance of <see cref="OptimizeRouteCriteria"/>
+        /// </summary>
         public OptimizeRouteCriteria(string startAddress, string endAddress, List<OptimizeRouteWaypoint> waypoints)
         {
             if (string.IsNullOrWhiteSpace(startAddress))
@@ -31,21 +37,19 @@ namespace AllReady.Services.Routing
             Waypoints = waypoints;
         }
 
+        /// <summary>
+        /// The start address for the optmized route
+        /// </summary>
         public string StartAddress { get; private set; }
+
+        /// <summary>
+        /// The end address for the optmized route
+        /// </summary>
         public string EndAddress { get; private set; }
+
+        /// <summary>
+        /// A list of waypoints to be optimized for the route
+        /// </summary>
         public List<OptimizeRouteWaypoint> Waypoints { get; private set; }
-    }
-
-    public class OptimizeRouteWaypoint
-    {
-        public OptimizeRouteWaypoint(double longitude, double latitude, Guid requestId)
-        {
-            Coordinates = string.Join(",", latitude.ToString(), longitude.ToString());
-            RequestId = requestId;
-        }
-
-        public string Coordinates { get; }
-
-        public Guid RequestId { get; }
     }
 }
