--- AllReadyApp/Web-App/AllReady/Services/Routing/IOptimizeRouteService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace AllReady.Services.Routing
-{
-    public interface IOptimizeRouteService
-    {
-        Task<OptimizeRouteResult> OptimizeRoute(OptimizeRouteCriteria criteria);
-    }
-}
