+++ AllReadyApp/Web-App/AllReady/Startup.cs
@@ -33,8 +33,9 @@ using Hangfire;
 using Hangfire.SqlServer;
 using AllReady.ModelBinding;
 using AllReady.Services.Mapping;
+using AllReady.Services.Mapping.GeoCoding;
+using AllReady.Services.Mapping.Routing;
 using Microsoft.AspNetCore.Localization;
-using AllReady.Services.Routing;
 using AllReady.Services.Twitter;
 using CsvHelper;
 using AllReady.Services.Sms;
